packages/scrapers/pyproject.toml
[tool.poetry]
name = "scrapers"
version = "0.1.0"
description = ""
authors = ["Mike Russo <russomichaelb@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
youtube_dl = "^2021.12.17"
webvtt-py = "^0.4.6"
subtitle-parser = "^1.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
